+++ lib/ect.rb
@@ -4,3 +4,8 @@ module Ect
   VERSION = '1.0.0'
 end
 
+module Enumerable
+
+  alias bisect partition
+end
+