+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 
 ## ect 1.0.0  not yet released
 
+- alias #bisect to Enumerable#partition
+